requirements.txt
black
coverage
django
django-environ
pytest
pytest-cov
pytest-django
pytest-mock
